Moved server events to their own class.
Moved BukkitMCServerCommandEvent and BukkitMCServerPingEvent from BukkitMiscEvents to BukkitServerEvents.
